+++ controllers/getEmployees.js
@@ -9,4 +9,23 @@ const getEmployees = (req, res) => {
   })
 }
 
-module.exports {getEmployees}
+const getEmployeesByFirstName = (req, res) => {
+  pool.query("SELECT first_name FROM employees", (err, rows) => {
+    if (err) return handleSQLError(res, err)
+    return res.json(rows);
+  })
+}
+
+const getEmployeesById = (req, res) => {
+  let sql = "SELECT * FROM users WHERE id = ?"
+
+  const id = req.params.id
+  sql = mysql.format(sql, [id])
+
+  pool.query(sql, (err, rows) => {
+    if (err) return handleSQLError(res, err)
+    return res.json(rows);
+  })
+}
+
+module.exports = {getEmployees, getEmployeesById, getEmployeesByFirstName}
