+++ index.js
@@ -1,7 +1,6 @@
 const express = require('express');
 const app = express();
 const port = process.env.PORT || 4001
-const employees = require('./routes/employees')
 
 app.get('/', (req, res) => res.send('WELCOME TO OUR API'))
